requirements.txt
asgiref==3.3.1; python_version >= "3.6"
django==3.1.7; python_version >= "3.6"
djangorestframework==3.12.2; python_version >= "3.5"
future==0.18.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
pytz==2021.1; python_version >= "3.6"
serializers==0.2.4
sqlparse==0.4.1; python_version >= "3.6"
